fix accidentally moved multi-use components
