+++ web/src/lib/components/AdminCard.svelte
