+++ web/src/lib/components/layouts/AdminPageLayout.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import BreadcrumbActionPage from './BreadcrumbActionPage.svelte';
+  import BreadcrumbActionPage from '$lib/components/BreadcrumbActionPage.svelte';
   import NavigationBar from '$lib/components/shared-components/navigation-bar/navigation-bar.svelte';
   import BottomInfo from '$lib/components/shared-components/side-bar/bottom-info.svelte';
   import { Route } from '$lib/route';
