+++ web/src/lib/components/server-statistics/ServerStatisticsCard.svelte
