+++ web/src/routes/admin/maintenance/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import AdminPageLayout from '../AdminPageLayout.svelte';
+  import AdminPageLayout from '$lib/components/layouts/AdminPageLayout.svelte';
   import MaintenanceBackupsList from '$lib/components/maintenance/MaintenanceBackupsList.svelte';
   import SettingAccordionState from '$lib/components/shared-components/settings/setting-accordion-state.svelte';
   import SettingAccordion from '$lib/components/shared-components/settings/setting-accordion.svelte';
