+++ web/src/routes/admin/queues/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import AdminPageLayout from '../AdminPageLayout.svelte';
+  import AdminPageLayout from '$lib/components/layouts/AdminPageLayout.svelte';
   import OnEvents from '$lib/components/OnEvents.svelte';
   import JobsPanel from './QueuePanel.svelte';
   import { queueManager } from '$lib/managers/queue-manager.svelte';
