+++ web/src/routes/admin/server-status/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import AdminPageLayout from '../AdminPageLayout.svelte';
+  import AdminPageLayout from '$lib/components/layouts/AdminPageLayout.svelte';
   import ServerStatisticsPanel from './ServerStatisticsPanel.svelte';
   import { getServerStatistics, type ServerStatsResponseDto } from '@immich/sdk';
   import { Container } from '@immich/ui';
