+++ web/src/routes/admin/server-status/ServerStatisticsPanel.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import ServerStatisticsCard from '../ServerStatisticsCard.svelte';
+  import ServerStatisticsCard from '$lib/components/server-statistics/ServerStatisticsCard.svelte';
   import { locale } from '$lib/stores/preferences.store';
   import { getBytesWithUnit } from '$lib/utils/byte-units';
   import type { ServerStatsResponseDto, UserAdminResponseDto } from '@immich/sdk';
